requirements.txt
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
openai==1.51.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<1.6.0
pgvector==0.3.3
pdfplumber==0.11.0
python-docx==1.1.2
langchain-text-splitters==0.3.0
httpx==0.27.2
celery==5.4.0
redis==5.0.1


